Constant HEADER_SIZE across processor architectures

+++ lib/src/blob/core.rs
@@ -1,7 +1,9 @@
 use crate::block::BLOCK_SIZE;
 use std::mem;
 
-pub(super) const HEADER_SIZE: usize = mem::size_of::<usize>();
+// Using u64 instead of usize because HEADER_SIZE must be the same irrespective of whether we're on
+// a 32bit or 64bit processor (if we want two such replicas to be able to sync).
+pub(super) const HEADER_SIZE: usize = mem::size_of::<u64>();
 
 #[derive(Debug)]
 pub(crate) struct Core {
